refactor: 핸들러 매핑 인터페이스 추출
새로운 핸들러 매퍼를 추가하기 전 공통적으로 사용할 메서드를 인터페이스로 추출함.
